not work if link_file empty

+++ templates/default/screengun.sh
@@ -1,21 +1,31 @@
 #!/bin/bash
 
-#server with a specific DISPLAY
-vncserver :11 -geometry 1280x2556
+LINK_FILE=url_list2.txt
 
-## read URLs from a data file in a loop
-while read url name
-do
-    export DISPLAY=:11
-    google-chrome --window-size=1282,2056 --window-position=0,0 --user-data-dir=stage-profile --kiosk --app=$url > /dev/null 2> /dev/null &
+if [ -s $LINK_FILE ]  
+then  
+   # echo "The File has some data." 
 
-    sleep 3
-    import -display :11 -window root ../../screenshots/$name
+  #server with a specific DISPLAY
+  vncserver :11 -geometry 1280x2556
+  
+  ## read URLs from a data file in a loop
+  while read url name
+  do
+      export DISPLAY=:11
+      google-chrome --window-size=1282,2056 --window-position=0,0 --user-data-dir=stage-profile --kiosk --app=$url > /dev/null 2> /dev/null &
+  
+      sleep 3
+      import -display :11 -window root ../../screenshots/$name
+  
+      killall chrome
+      sleep 1
+      
+  done < $LINK_FILE
+  
+  ## clean up when done
+  vncserver -kill :11
 
-    killall chrome
-    sleep 1
-    
-done < url_list.txt
-
-## clean up when done
-vncserver -kill :11
+else  
+    echo "The File is empty."
+fi
